LayoutTests/fast/canvas/imageData-consistency-expected.txt

PASS actual.width is getData.width
PASS actual.height is getData.height
PASS actual.width is getData.width
PASS actual.height is getData.height
PASS actual.width is getData.width
PASS actual.height is getData.height
PASS actual.width is getData.width
PASS actual.height is getData.height
PASS actual.width is getData.width
PASS actual.height is getData.height
PASS actual.width is getData.width
PASS actual.height is getData.height
PASS actual.width is getData.width
PASS actual.height is getData.height
PASS actual.width is getData.width
PASS actual.height is getData.height
PASS actual.width is getData.width
PASS actual.height is getData.height
PASS actual.width is getData.width
PASS actual.height is getData.height
PASS actual.width is getData.width
PASS actual.height is getData.height
PASS actual.width is getData.width
PASS actual.height is getData.height
PASS actual.width is getData.width
PASS actual.height is getData.height
PASS actual.width is getData.width
PASS actual.height is getData.height
PASS actual.width is getData.width
PASS actual.height is getData.height
PASS successfullyParsed is true

TEST COMPLETE
